Added restarting options

+++ src/PyRsw.py
@@ -53,7 +53,7 @@ class Simulation:
         self.f0   = 1e-4            # Coriolis
         self.beta = 0.
         self.cfl  = 0.5             # default CFL
-        self.time = 0               # initial time
+        self.time = 0.              # initial time
         self.min_dt = 1e-3          # minimum timestep
 
         self.geomx = 'periodic'     # x boundary condition
@@ -75,9 +75,13 @@ class Simulation:
 
         self.plott = np.inf
         self.diagt = np.inf
+        self.savet = np.inf
 
         self.num_steps = 0
         self.mean_dt = 0.
+
+        self.restarting = False
+        self.restart_index = 0
 
         self.topo_func = null_topo  # Default to no topograpy
         
@@ -101,6 +105,8 @@ class Simulation:
                 print('Coriolis = beta-plane')
             else:
                 print('Coriolis = f-plane')
+        if self.restarting:
+            print('Restarting from index {0:d}'.format(self.restart_index))
         print ' '
         
         # Initialize grids and cell centres
@@ -170,6 +176,18 @@ class Simulation:
                 
         self.sfilt = np.tile(filtx,(1,self.Nky))*np.tile(filty,(self.Nkx,1))
             
+        # If we're restarting load the appropriate files
+        if self.restarting:
+            try:
+                restart_file = np.load('Outputs/{0:s}/{1:05d}.npz'.format(self.run_name,self.restart_index))
+                self.soln.u = restart_file['u']
+                self.soln.v = restart_file['v']
+                self.soln.h = restart_file['h']
+                self.time = restart_file['t']
+            except:
+                print('Restarting failed. Exiting.')
+                sys.exit()
+
         
     def prepare_for_run(self):
 
@@ -178,7 +196,8 @@ class Simulation:
     
         # If we're saving, initialize those too
         if self.output:
-            self.next_save_time = self.savet
+            self.next_save_time = (self.restart_index+1)*self.savet
+            self.out_counter = self.restart_index
     
         # If we're saving, initialize the directory
         if self.output or (self.animate == 'Save') or self.diagnose:
@@ -203,8 +222,10 @@ class Simulation:
             self.hov_count = 1
 
             self.initialize_plots(self)
-            self.update_plots(self)
-            self.next_plot_time = self.plott
+            if not(self.restarting):
+                self.update_plots(self)
+            self.frame_count = int(np.floor(self.time/self.plott) + 1)
+            self.next_plot_time = self.frame_count*self.plott
 
     # Compute the current flux
     def flux(self):
@@ -353,20 +374,21 @@ class Simulation:
     # Initialize the saving
     def initialize_saving(self):
         
-        path = 'Outputs/{0:s}'.format(self.run_name)
-        if not(os.path.isdir('Outputs')):
-            os.mkdir('Outputs')
+        if not(self.restarting):
+            path = 'Outputs/{0:s}'.format(self.run_name)
+            if not(os.path.isdir('Outputs')):
+                os.mkdir('Outputs')
 
-        # If directory already exists, delete it.
-        if os.path.isdir(path):
-           print('Output directory {0:s} already exists. '.format(path) + \
-                 'Warning, deleting everything in the directory.') 
-           shutil.rmtree(path)
-        # Make directory.
-        os.mkdir(path)
+            # If directory already exists, delete it.
+            if os.path.isdir(path):
+               print('Output directory {0:s} already exists. '.format(path) + \
+                     'Warning, deleting everything in the directory.') 
+               shutil.rmtree(path)
+            # Make directory.
+            os.mkdir(path)
 
-        if self.animate == 'Save':
-            os.mkdir(path+'/Frames')
+            if self.animate == 'Save':
+                os.mkdir(path+'/Frames')
 
         # Initialize saving stuff
         self.save_info()
